feat: change to architecture 1.0.0 for Crosstalent
