+++ src/hrflow_connectors/__init__.py
@@ -14,3 +14,6 @@ except PackageNotFoundError:  # pragma: no cover
     __version__ = "unknown"
 finally:
     del version, PackageNotFoundError
+
+from .connectors import *
+from .core.auth import *
