+++ src/hrflow_connectors/connectors/__init__.py
@@ -1,0 +1,1 @@
+from .crosstalent import *
