--- src/hrflow_connectors/connectors/boards/crosstalent/spec.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-from typing import Tuple
-
-from .actions import CrosstalentPullJobsAction
-
-
-class Spec(BaseModel):
-    actions: Tuple[CrosstalentPullJobsAction]
