+++ src/hrflow_connectors/connectors/crosstalent/README.md
@@ -1,0 +1,6 @@
+# Crosstalent Connector
+**Crosstalent designs and implements innovative solutions for human resources management.**
+
+Lists of actions :
+* [**Pull jobs**](docs/pull_jobs.md)
+* [**Push profile**](docs/push_profile.md)
