+++ src/hrflow_connectors/connectors/crosstalent/__init__.py
@@ -1,0 +1,1 @@
+from .connector import *
