+++ src/hrflow_connectors/connectors/crosstalent/docs/push_profile.md
@@ -1,12 +1,9 @@
-# Crosstalent Connector
-**Crosstalent designs and implements innovative solutions for human resources management.**
-
+# Push profile
 `Hrflow.ai` :arrow_right: `Crosstalent`
 
-## PushProfile
-`PushProfile` pushes a Hrflow.ai profile to `Crosstalent` via their ***Salesforce API***.
+`PushProfileAction` pushes a Hrflow.ai profile to `Crosstalent` via their ***Salesforce API***.
 
-### Parameters
+## Parameters
 
 | Field | Type | Description |
 | ----- | ---- | ----------- |
@@ -19,16 +16,16 @@
 | `auth` :red_circle: | `OAuth2PasswordCredentialsBody` | Auth instance to identify and communicate with the platform        |
 | `subdomain` :red_circle: | `str` | Subdomain Crosstalent just before `salesforce.com`. For example subdomain=`my_subdomain.my` in `http://my_subdomain.my.salesforce.com/ABC`        |
 
-:red_circle: : *required* 
+:red_circle: : *required*
 
-### Example
+## Example
 Let's take as an example in a [***CATCH workflow***](https://developers.hrflow.ai/docs/workflows#catch-setup).
 ```python
-from hrflow import Hrflow
+from hrflow_connectors import Crosstalent
 
-from hrflow_connectors.core.auth import OAuth2PasswordCredentialsBody
-from hrflow_connectors.connectors.destinations.crosstalent import PushProfile
-from hrflow_connectors.utils.hrflow import EventParser, Profile, Source
+from hrflow import Hrflow
+from hrflow_connectors import OAuth2PasswordCredentialsBody
+from hrflow_connectors.utils.hrflow import Profile
 from hrflow_connectors.utils.logger import get_logger_with_basic_config
 
 
@@ -48,21 +45,19 @@ def workflow(_request, settings):
 
         client = Hrflow(api_secret=settings["X-API-KEY"], api_user=settings["X-USER-EMAIL"])
 
-        access_token_url = "https://test.salesforce.com/services/oauth2/token"
         auth = OAuth2PasswordCredentialsBody(
-            access_token_url=access_token_url,
+            access_token_url="https://test.salesforce.com/services/oauth2/token",
             client_id=settings["CLIENT_ID"],
             client_secret=settings["CLIENT_SECRET"],
             username=settings["USERNAME"],
             password=settings["PASSWORD"],
         )
 
-        action = PushProfile(
+        response = Crosstalent.push_profile(
             auth=auth,
             subdomain=settings["SUBDOMAIN"],
             hrflow_client=client,
             profile=profile,
         )
-        response = action.execute()
         return response
 ```
