--- src/hrflow_connectors/connectors/destinations/crosstalent/__init__.py
@@ -1,2 +1,0 @@
-from .actions import *
-from .spec import *
