--- src/hrflow_connectors/connectors/destinations/crosstalent/spec.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-from typing import Tuple
-
-from .actions import CrosstalentPushProfileAction
-
-
-class Spec(BaseModel):
-    actions: Tuple[CrosstalentPushProfileAction]
