+++ src/hrflow_connectors/core/connector.py
@@ -1,30 +1,34 @@
+from typing import Optional, Dict, Any
+
+
 class Connector:
     """
     Abstract class `Connector`
     """
+
     @staticmethod
-    def pull_jobs(*args, **kwargs):
+    def pull_jobs(*args, **kwargs) -> Optional[Dict[str, Any]]:
         """
         Pull jobs
         """
         raise NotImplementedError("`pull_jobs` method is not implemented !")
 
     @staticmethod
-    def pull_profiles(*args, **kwargs):
+    def pull_profiles(*args, **kwargs) -> Optional[Dict[str, Any]]:
         """
         Pull profiles
         """
         raise NotImplementedError("`pull_profiles` method is not implemented !")
 
     @staticmethod
-    def push_job(*args, **kwargs):
+    def push_job(*args, **kwargs) -> Optional[Dict[str, Any]]:
         """
         Push job
         """
         raise NotImplementedError("`push_job` method is not implemented !")
 
     @staticmethod
-    def push_profile(*args, **kwargs):
+    def push_profile(*args, **kwargs) -> Optional[Dict[str, Any]]:
         """
         Push profile
         """
