+++ tests/connectors/crosstalent/test_pull_jobs.py
@@ -1,10 +1,7 @@
 import pytest
 
 from hrflow_connectors.core.auth import OAuth2PasswordCredentialsBody
-from hrflow_connectors.connectors.destinations.crosstalent import (
-    CrosstalentPushProfileAction,
-)
-from hrflow_connectors.utils.hrflow import Profile, Source
+from hrflow_connectors.connectors.crosstalent import PullJobsAction
 
 
 @pytest.fixture
@@ -21,16 +18,18 @@ def auth(credentials):
     return auth
 
 
-def test_CrosstalentPushProfileAction(logger, auth, hrflow_client):
-    profile = Profile(
-        key="ea5704b959c5e53aaef65c04ef5018ae1fee1a77",
-        source=Source(key="15517d70b0870e4cf431eefd78f8b39cff5607e8"),
-    )
-    action = CrosstalentPushProfileAction(
+def test_Auth(auth):
+    access_token = auth.get_access_token()
+    assert isinstance(access_token, str)
+    assert access_token != ""
+
+
+def test_PullJobsAction(logger, auth, hrflow_client):
+    action = PullJobsAction(
         auth=auth,
         subdomain="vulcain-eng--recette.my",
-        hrflow_client=hrflow_client("vulcain"),
-        profile=profile,
+        hrflow_client=hrflow_client("dev-demo"),
+        board_key="8eba188e1af123a9818d00974ff37b943b7d54f4",
+        hydrate_with_parsing=True,
     )
-    response = action.execute()
-    assert response.get("status_code") == 201
+    action.execute()
