+++ tests/connectors/crosstalent/test_push_profile.py
@@ -1,7 +1,8 @@
 import pytest
 
 from hrflow_connectors.core.auth import OAuth2PasswordCredentialsBody
-from hrflow_connectors.connectors.boards.crosstalent import CrosstalentPullJobsAction
+from hrflow_connectors.connectors.crosstalent import PushProfileAction
+from hrflow_connectors.utils.hrflow import Profile, Source
 
 
 @pytest.fixture
@@ -18,18 +19,16 @@ def auth(credentials):
     return auth
 
 
-def test_Auth(auth):
-    access_token = auth.get_access_token()
-    assert isinstance(access_token, str)
-    assert access_token != ""
-
-
-def test_GetAllJobs(logger, auth, hrflow_client):
-    action = CrosstalentPullJobsAction(
+def test_PushProfileAction(logger, auth, hrflow_client):
+    profile = Profile(
+        key="ea5704b959c5e53aaef65c04ef5018ae1fee1a77",
+        source=Source(key="15517d70b0870e4cf431eefd78f8b39cff5607e8"),
+    )
+    action = PushProfileAction(
         auth=auth,
         subdomain="vulcain-eng--recette.my",
-        hrflow_client=hrflow_client("dev-demo"),
-        board_key="8eba188e1af123a9818d00974ff37b943b7d54f4",
-        hydrate_with_parsing=True,
+        hrflow_client=hrflow_client("vulcain"),
+        profile=profile,
     )
-    action.execute()
+    response = action.execute()
+    assert response.get("status_code") == 201
